Refactored layers.py
Layer is now an abstract class, with different Layer instances being defined in this class. This should make it easier to extend in the future. I am still not sure how much modelling of the layers I want to be done by me, or even if this would be useful when PyTorch and TensorFlow compatability is added.